refactor(todo): :recycle: give its own file to Column
