+++ client/src/components/containers/index.js
@@ -2,6 +2,7 @@ import RegisterForm from './RegisterForm';
 import LoginForm from './LoginForm';
 import AddBoard from './AddBoard';
 import ButtonAddBoard from './ButtonAddBoard';
-import Todo from "./Todo"
+import Todo from './Todo';
+import Column from './Column';
 
-export { RegisterForm, AddBoard, ButtonAddBoard, LoginForm, Todo };
+export { RegisterForm, AddBoard, ButtonAddBoard, LoginForm, Todo, Column };
