+++ client/src/styles/containers/todo.css
@@ -12,21 +12,6 @@
   overflow-x: auto;
   overflow-y: hidden;
 }
-.column {
-  display: grid;
-  grid-template-columns: minmax(260px, 1fr);
-  grid-template-rows: 1fr 11fr;
-  height: 73vh;
-  justify-content: center;
-}
-.taks {
-  display: grid;
-  grid-template-columns: minmax(250px, 1fr);
-  grid-template-rows: repeat(auto-fill, minmax(148px, 1fr));
-  grid-gap: 2%;
-  overflow-x: hidden;
-  overflow-y: auto;
-}
 .taks::-webkit-scrollbar,
 .container-columns::-webkit-scrollbar {
   width: 10px;
